Add support for Svelte.js
Hi @vunguyentuan!

Thanks for working on the extension!

From the looks of it - this seems to be all that's needed to add Svelte support. If you don't mind, I'd appreciate something like this being merged in the next version. I've monkeypatched the extension at the moment for myself, but if an update is released without this patch I'll have to monkeypatch again...and again 🙂